// server/src/models/Card/CardKey.ts
export enum CardKey {
  id = "id",
  title = "title",
  idUser = "idUser",
  idList = "idList",
  idBoard = "idBoard",
  dateDue = "dateDue",
  idLabels = "idLabels",
  dateStart = "dateStart",
  isWatched = "isWatched",
  isComplete = "isComplete",
  description = "description",
  dateCreation = "dateCreation",
  dateLastView = "dateLastView",
  idChecklists = "idChecklists",  
}


export default CardKey